Distance and time calculation for onroad transport .py
The project calculates time and distance or supply chain route for on-road transport using an API service.
For calculating mobility, it is required that Python be used in relation to create our database by adding our locations (latitude
longitudes) from destination to source into that Python code framework of calculating distance and time. For calculation purposes, one has to observe how much the distance and the time it takes for material to be transported on roads. The study gives emphasis to the creation of a database for the calculation of supply chain data to anticipate the cost and logistic calculations.
The panda is used to store the data by creating two dimensional data frames; hence, the spreadsheet can be created, to find a library to store the data of location and time it requires to
reach material from one place to another by fetching the data of latitude and longitude from our
database and making requests to API server to find our Geo statistical data( Martin et al., 2023).
It may occur that the supply chains are sometimes part of circular economy or dependent
economy which means that one supply chain depends on further supply chain and next supply
chain in the journey of material transfer, any contingencies in the previous supply chain can affect the
next supply chain negatively or positively, the performance of previous supply chain is positive
And then it can avoid positive output towards the next supply chain and if the supply chain. If the performance of previous supply chain is negative, it can affect overall economy of next supply chain in terms of cost of material demands and inventory storage and the cost related to inventory storage—snap—hence increasing overall cost of supply chain and decreasing the efficiency of supply chain